Preserve FACTOR
I wrote the first version of this program the summer after writing
QUADFORM. I worked on it on and off for a few months, often drawing out
pictures of the "box method" for factoring a quadratic equation to
guide my design. Those drawings constituted the first time I had truly
solved the general case of a math problem, which certainly prepared me
for learning proofs in geometry the following year (though I didn't
recognize it at the time).

When the program was finally done, I felt so proud of myself! This
particular program won me lots of attention from fellow classmates and
even from a few future teachers. It likely saved me enough time on
homework and exams to compensate for the time I invested into writing
it.

Since I used this program so much, I gradually added additional
enhancements. For example, when factoring an equation with large
coefficients, the factoring computation got pretty slow, so I added in
a "CALCULATING..." animation so I would know that the program hadn't
frozen.* Later, I added a few extra checks like the a+bi mode and
computing the discriminant to further improve performance and the user
experience.

* I now know that the long computation times for finding factors stem
from an inherent complexity in that problem. In fact, it is that very
computation difficulty which makes prime number factorization so
desirable for most modern encryption techniques.