Preserve LOADING
This program clearly has some parallels to middle block of FACTOR
that shows a "CALCULATING..." animation when the numbers to factor are
large.

I know that the first version of FACTOR definitely preceded LOADING,
but I don't remember if I added the "CALCULATING..." animation before
or after writing this program.

As a goofy high school student, I had fun pranking my friends with this
program. When they weren't looking, I used a link cable to copy this
program to their calculator and start it up. When they got back, they
just saw their calculator in a long LOADING animation. Even better,
since the TI calculators have a slow clock speed, this program would
run for a long time, causing my friends to get super frustrated that
their calculator was "unusable" for so long! (Of course I didn't tell
them that pressing the ON button would kill the program).