Preserve SURPRISE
This is a less nefarious version of the LOADING program that I wrote
later in my high school career. Like LOADING, I would find a moment on
a friend's birthday to steal his/her calculator and copy the program
over. I'd then usally open the PRGM menu so that the next thing my
friend saw was the word SURPRISE on their screen. I always had fun
seeing their smiles as they saw the HAPPY BIRTHDAY! animation. Even
better was the fact that they immediately knew it was from me since I
was the only one that knew how to program the calculator :D